+++ src/layers/MultiscaleImageLayer/MultiscaleImageLayer.js
@@ -1,17 +1,29 @@
 import { CompositeLayer } from '@deck.gl/core';
 import { isWebGL2 } from '@luma.gl/core';
 
-import VivViewerLayerBase from './VivViewerLayerBase';
-import StaticImageLayer from '../StaticImageLayer';
-import { to32BitFloat, getNearestPowerOf2 } from '../utils';
+import MultiscaleImageLayerBase from './MultiscaleImageLayerBase';
+import ImageLayer from '../ImageLayer';
+import { to32BitFloat, getNearestPowerOf2, onPointer } from '../utils';
 
 const defaultProps = {
   pickable: true,
-  onHover: { type: 'function', value: null, compare: false }
+  onHover: { type: 'function', value: null, compare: false },
+  sliderValues: { type: 'array', value: [], compare: true },
+  colorValues: { type: 'array', value: [], compare: true },
+  channelIsOn: { type: 'array', value: [], compare: true },
+  opacity: { type: 'number', value: 1, compare: true },
+  colormap: { type: 'string', value: '', compare: true },
+  domain: { type: 'array', value: [], compare: true },
+  viewportId: { type: 'string', value: '', compare: true },
+  isLensOn: { type: 'boolean', value: false, compare: true },
+  lensSelection: { type: 'number', value: 0, compare: true },
+  lensRadius: { type: 'number', value: 100, compare: true },
+  lensBorderColor: { type: 'array', value: [255, 255, 255], compare: true },
+  lensBorderRadius: { type: 'number', value: 0.02, compare: true }
 };
 
 /**
- * This layer generates a VivViewerLayer (tiled) and a StaticImageLayer (background for the tiled layer)
+ * This layer generates a MultiscaleImageLayer (tiled) and a ImageLayer (background for the tiled layer)
  * @param {Object} props
  * @param {Array} props.sliderValues List of [begin, end] values to control each channel's ramp function.
  * @param {Array} props.colorValues List of [r, g, b] values for each channel.
@@ -25,9 +37,27 @@ const defaultProps = {
  * @param {String} props.id Unique identifier for this layer.
  * @param {String} props.onTileError Custom override for handle tile fetching errors.
  * @param {String} props.onHover Hook function from deck.gl to handle hover objects.
+ * @param {boolean} props.isLensOn Whether or not to use the lens.
+ * @param {number} props.lensSelection Numeric index of the channel to be focused on by the lens.
+ * @param {number} props.lensRadius Pixel radius of the lens (default: 100).
+ * @param {number} props.lensBorderColor RGB color of the border of the lens (default [255, 255, 255]).
+ * @param {number} props.lensBorderRadius Percentage of the radius of the lens for a border (default 0.02).
  */
 
-export default class VivViewerLayer extends CompositeLayer {
+export default class MultiscaleImageLayer extends CompositeLayer {
+  initializeState() {
+    this.state = {
+      unprojectLensBounds: [0, 0, 0, 0]
+    };
+    if (this.context.deck) {
+      this.context.deck.eventManager.on({
+        pointermove: () => onPointer(this),
+        pointerleave: () => onPointer(this),
+        wheel: () => onPointer(this)
+      });
+    }
+  }
+
   renderLayers() {
     const {
       loader,
@@ -42,9 +72,14 @@ export default class VivViewerLayer extends CompositeLayer {
       onTileError,
       onHover,
       pickable,
-      id
+      id,
+      isLensOn,
+      lensSelection,
+      lensBorderColor,
+      lensBorderRadius
     } = this.props;
     const { tileSize, numLevels, dtype } = loader;
+    const { unprojectLensBounds } = this.state;
     const noWebGl2 = !isWebGL2(this.context.gl);
     const getTileData = async ({ x, y, z }) => {
       const tile = await loader.getTile({
@@ -70,7 +105,7 @@ export default class VivViewerLayer extends CompositeLayer {
       return tile;
     };
     const { height, width } = loader.getRasterSize({ z: 0 });
-    const tiledLayer = new VivViewerLayerBase({
+    const tiledLayer = new MultiscaleImageLayerBase({
       id: `Tiled-Image-${id}`,
       getTileData,
       dtype,
@@ -96,7 +131,12 @@ export default class VivViewerLayer extends CompositeLayer {
       colormap,
       viewportId,
       onHover,
-      pickable
+      pickable,
+      unprojectLensBounds,
+      isLensOn,
+      lensSelection,
+      lensBorderColor,
+      lensBorderRadius
     });
     // This gives us a background image and also solves the current
     // minZoom funny business.  We don't use it for the background if we have an opacity
@@ -108,7 +148,7 @@ export default class VivViewerLayer extends CompositeLayer {
     });
     const baseLayer =
       implementsGetRaster &&
-      new StaticImageLayer(this.props, {
+      new ImageLayer(this.props, {
         id: `Background-Image-${id}`,
         scale: 2 ** (numLevels - 1),
         visible:
@@ -125,5 +165,5 @@ export default class VivViewerLayer extends CompositeLayer {
   }
 }
 
-VivViewerLayer.layerName = 'VivViewerLayer';
-VivViewerLayer.defaultProps = defaultProps;
+MultiscaleImageLayer.layerName = 'MultiscaleImageLayer';
+MultiscaleImageLayer.defaultProps = defaultProps;