+++ src/layers/MultiscaleImageLayer/index.js
@@ -1,0 +1,3 @@
+import MultiscaleImageLayer from './MultiscaleImageLayer';
+
+export default MultiscaleImageLayer;