+++ src/layers/VivViewerLayer/index.js
@@ -1,3 +1,0 @@
-import VivViewerLayer from './VivViewerLayer';
-
-export default VivViewerLayer;