+++ src/layers/index.js
@@ -1,5 +1,5 @@
-export { default as VivViewerLayer } from './VivViewerLayer';
-export { default as StaticImageLayer } from './StaticImageLayer';
+export { default as MultiscaleImageLayer } from './MultiscaleImageLayer';
+export { default as ImageLayer } from './ImageLayer';
 export { default as OverviewLayer } from './OverviewLayer';
 export { default as ScaleBarLayer } from './ScaleBarLayer';
 export { default as XRLayer } from './XRLayer';