+++ src/loaders/index.js
@@ -1,10 +1,10 @@
-import { openArray } from 'zarr';
+import { openArray, HTTPStore } from 'zarr';
 import { fromUrl } from 'geotiff';
 import Pool from './Pool';
 import ZarrLoader from './zarrLoader';
 import OMETiffLoader from './OMETiffLoader';
-import { getChannelStats } from './utils';
-import OMEZarrReader from './omeZarrReader';
+import { getChannelStats, getJson, dimensionsFromOMEXML } from './utils';
+import OMEXML from './omeXML';
 
 export async function createZarrLoader({
   url,
@@ -14,6 +14,9 @@ export async function createZarrLoader({
   scale,
   translate
 }) {
+  // TODO: This is a legacy initialization function. There is an official
+  // specification now for multiscale datasets (see below), that doesn't
+  // consolidate metadata in this way.
   let data;
   if (isPyramid) {
     const metadataUrl = `${url}${url.slice(-1) === '/' ? '' : '/'}.zmetadata`;
@@ -33,6 +36,68 @@ export async function createZarrLoader({
     translate,
     isRgb
   });
+}
+
+export async function createBioformatsZarrLoader({ url }) {
+  const baseUrl = url.endsWith('/') ? url : `${url}/`;
+  const metaUrl = `${baseUrl}METADATA.ome.xml`;
+  const store = new HTTPStore(`${baseUrl}data.zarr/0`); // first image
+  const rootAttrs = await getJson(store, '.zattrs');
+
+  let resolutions = ['0'];
+  if ('multiscales' in rootAttrs) {
+    // Get path to subresolutions if they exist
+    const { datasets } = rootAttrs.multiscales[0];
+    resolutions = datasets.map(d => d.path);
+  }
+
+  const promises = resolutions.map(path => openArray({ store, path }));
+  const pyramid = await Promise.all(promises);
+
+  /*
+   * TODO: There should be a much better way to do this.
+   * If base image is small, we don't need to fetch data for the
+   * top levels of the pyramid. For large images, the tile sizes (chunks)
+   * will be the same size for x/y. We check the chunksize here for this edge case.
+   */
+  const { chunks, shape } = pyramid[0];
+  const shouldUseBase = chunks[-1] !== chunks[-2];
+  const data = pyramid.length > 1 || shouldUseBase ? pyramid : pyramid[0];
+
+  // Get OMEXML string
+  const buffer = await fetch(metaUrl).then(res => res.arrayBuffer());
+  const omexmlString = new TextDecoder().decode(new Uint8Array(buffer));
+  const omexml = new OMEXML(omexmlString);
+  const dimensions = dimensionsFromOMEXML(omexml);
+
+  /*
+   * Specifying different dimension orders form the METADATA.ome.xml is
+   * possible and necessary for creating an OME-Zarr precursor.
+   *
+   * e.g. `bioformats2raw --file_type=zarr --dimension-order='XYZCY'`
+   *
+   * Here we check the shape of base of the pyrmaid and compare the shape
+   * to the shape of the dimensions. If they are different, we reorder the
+   * dimensions to create the zarr loader. This is fragile code, and will only
+   * be executed if someone tries to specify different dimension orders.
+   */
+  const nonXYShape = shape.slice(0, -2); // XY always last dims and don't need to be compared
+  const nonXYDims = dimensions.filter(d => d.values); // XY are null
+  const allSameSize = nonXYShape.every(
+    (s, i) => s === nonXYDims[i].values.length
+  );
+  if (!allSameSize) {
+    const sortedDims = [];
+    // Greedily match first matching dimension
+    nonXYShape.forEach(len => {
+      const firstMatchedDim = nonXYDims.filter(d => d.values.length === len)[0];
+      sortedDims.push(firstMatchedDim);
+    });
+    const newDimensions = [...sortedDims, ...dimensions.slice(-2)]; // append YX dims
+    return new ZarrLoader({ data, dimensions: newDimensions });
+  }
+
+  return new ZarrLoader({ data, dimensions });
 }
 
 /**
@@ -56,4 +121,4 @@ export async function createOMETiffLoader({ url, offsets = [], headers = {} }) {
   });
 }
 
-export { ZarrLoader, OMETiffLoader, OMEZarrReader, getChannelStats };
+export { ZarrLoader, OMETiffLoader, getChannelStats };